Revert "Get rid of exit module and move utility functions into main.rs"
This reverts commit d948775180ea1990cfe6e23f610bddb81e72c2aa.
